feat : 식별자 랜덤 생성, 방생성, 여러 게임 기능
