+++ src/main/java/wooteco/chess/domain/board/BoardFactory.java
@@ -10,6 +10,14 @@ import wooteco.chess.domain.position.Position;
 public class BoardFactory {
 	private static final Pattern PIECES_PATTERN = Pattern.compile("([\\.pPbBrRkKnNqQ]{8}\\n){8}");
 	private static final String ILLEGAL_BOARD_REGEX_EXCEPTION_MESSAGE = "문자열의 형태가 체스판의 형식이 아닙니다.";
+	private static final String EMPTY_BOARD = "........\n"
+		+ "........\n"
+		+ "........\n"
+		+ "........\n"
+		+ "........\n"
+		+ "........\n"
+		+ "........\n"
+		+ "........\n";
 
 	private BoardFactory() {
 	}
@@ -24,5 +32,9 @@ public class BoardFactory {
 		if (!PIECES_PATTERN.matcher(boards).matches()) {
 			throw new IllegalArgumentException(String.format(ILLEGAL_BOARD_REGEX_EXCEPTION_MESSAGE + "%s", boards));
 		}
+	}
+
+	public static Board ofEmpty() {
+		return create(EMPTY_BOARD);
 	}
 }
