+++ src/main/java/wooteco/chess/repository/jdbc/GameRepository.java
@@ -1,7 +1,12 @@
 package wooteco.chess.repository.jdbc;
 
+import java.util.List;
+
+import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GameRepository extends CrudRepository<GameEntity, Long> {
+public interface GameRepository extends CrudRepository<GameEntity, String> {
+	@Query("SELECT id FROM game")
+	List<String> findAllIds();
 
 }
