+++ src/main/resources/schema.sql
@@ -1,5 +1,5 @@
 create table if not exists game(
-    id bigint auto_increment,
+    id varchar(40),
     state varchar(255) not null,
     turn varchar(255) not null,
     board varchar(255) not null,
