+++ src/main/resources/templates/game.hbs
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+    <meta charset="UTF-8">
+    <title>CHESS_GAME</title>
+    <link rel="stylesheet" href="../css/index.css"/>
+</head>
+<body>
+<div id="total-board">
+    <div id="main-board">
+        <div id="chess-title">
+            우 아 한 체 스 코 스 <span id="game-id">{{id}}</span>
+        </div>
+        <div id="board-cover">
+            <table id="chess-board">
+                <tbody>
+                </tbody>
+            </table>
+        </div>
+        <div>현재턴 : <span id="current-turn"></span></div>
+        <div>게임상태 : <span id="current-state">  </span></div>
+    </div>
+    <div>
+        <div>
+            <div>점수</div>
+            <div>
+                <label>흑팀 :</label>
+                <span class="black-team score">   </span>
+                <label>백팀 :</label>
+                <span class="white-team score">   </span>
+            </div>
+        </div>
+        <button id="new-game">새게임</button>
+        <button id="end-game">게임종료</button>
+    </div>
+</div>
+</body>
+<script src="../js/game.js"></script>
+</html>
