+++ src/main/resources/templates/index.hbs
@@ -1,38 +1,17 @@
 <!DOCTYPE html>
-<html lang="ko">
+<html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>CHESS_GAME</title>
-    <link rel="stylesheet" href="./css/index.css"/>
+    <title>CHESS HOME</title>
 </head>
 <body>
-<div id="total-board">
-    <div id="main-board">
-        <div id="chess-title">
-            우 아 한 체 스 코 스
-        </div>
-        <div id="board-cover">
-            <table id="chess-board">
-                <tbody>
-                </tbody>
-            </table>
-        </div>
-        <div>현재턴 : <span id="current-turn"></span></div>
-        <div>게임상태 : <span id="current-state">  </span></div>
-    </div>
+<div>
     <div>
-        <div>
-            <div>점수</div>
-            <div>
-                <label>흑팀 :</label>
-                <span class="black-team score">   </span>
-                <label>백팀 :</label>
-                <span class="white-team score">   </span>
-            </div>
-        </div>
-        <button id="new-game">새게임</button>
-        <button id="end-game">게임종료</button>
+        <button id="add-game">새 게임 추가</button>
     </div>
+    <ul id="game-room">
+
+    </ul>
 </div>
 </body>
 <script src="./js/index.js"></script>
