+++ src/test/java/wooteco/chess/ChessApplicationTests.java
@@ -1,5 +1,7 @@
 package wooteco.chess;
 
+import java.util.UUID;
+
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -8,6 +10,8 @@ class ChessApplicationTests {
 
 	@Test
 	void contextLoads() {
+		UUID uuid = UUID.randomUUID();
+		System.out.println(uuid);
 	}
 
 }
